mark complete / incomplete

+++ client/pages/index.js
@@ -27,18 +27,30 @@ const Index = () => {
       })
   }
 
-  const deleteTask = task => {
-    axios.delete(taskUrl(task)).then(({ data }) => {
-      const filtered = tasks.filter(t => t.id !== data.task.id)
-      setTasks(filtered)
-      setFlash({ messages: ['Task successfully destroyed'] })
+  const updateTask = task => {
+    axios.patch(taskUrl(task)).then(({ data }) => {
+      const index = tasks.findIndex(t => t.id === task.id)
+      tasks[index] = data.task
+      setTasks(tasks)
+      setFlash({ messages: ['Task successfully updated'] })
     })
+  }
+
+  const deleteTask = task => {
+    if(confirm("Are you sure?")){
+      axios.delete(taskUrl(task)).then(({ data }) => {
+        const filtered = tasks.filter(t => t.id !== data.task.id)
+        setTasks(filtered)
+        setFlash({ messages: ['Task successfully deleted'] })
+      })
+    }
   }
 
   return(
     <React.Fragment>
       <Tasks
         tasks={tasks}
+        updateTask={updateTask}
         deleteTask={deleteTask}
       />
       <hr></hr>
